chore: organize diagnostics module
