+++ kubert/src/admin/diagnostics.rs
@@ -1,49 +1,20 @@
-use ahash::AHashMap;
-use k8s_openapi::apimachinery::pkg::apis::meta::v1::{ObjectMeta, Time};
-use kube_runtime::watcher;
-use parking_lot::{Mutex, RwLock};
-use std::{
-    borrow::Cow,
-    net::SocketAddr,
-    sync::{Arc, Weak},
-};
+use k8s_openapi::apimachinery::pkg::apis::meta::v1::Time;
+use parking_lot::Mutex;
+use std::{net::SocketAddr, sync::Arc};
+
+#[cfg(feature = "lease")]
+mod lease;
+mod watch;
+
+#[cfg(feature = "lease")]
+pub(crate) use self::lease::LeaseDiagnostics;
+pub(crate) use self::watch::WatchDiagnostics;
 
 #[derive(Clone, Debug)]
 pub(crate) struct Diagnostics {
     initial_time: chrono::DateTime<chrono::Utc>,
-    leases: Arc<Mutex<Vec<Weak<RwLock<LeaseState>>>>>,
-    watches: Arc<Mutex<Vec<Weak<RwLock<WatchState>>>>>,
-}
-
-pub(crate) struct WatchDiagnostics(Arc<RwLock<WatchState>>);
-
-#[cfg(feature = "lease")]
-pub(crate) struct LeaseDiagnostics(Arc<RwLock<LeaseState>>);
-
-#[derive(Clone, Debug)]
-struct WatchState {
-    api_url: String,
-    label_selector: String,
-    stats: WatchStats,
-    known: AHashMap<ObjRef, Resource>,
-    resetting: AHashMap<ObjRef, Resource>,
-}
-
-#[cfg(feature = "lease")]
-#[derive(Clone, Debug, serde::Serialize)]
-#[serde(rename_all = "camelCase")]
-struct LeaseState {
-    name: String,
-    namespace: String,
-    claimant: String,
-    lease_duration_seconds: f64,
-    renew_grace_period_seconds: f64,
-    field_manager: Cow<'static, str>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    claim: Option<crate::lease::Claim>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    resource_version: Option<String>,
-    stats: LeaseStats,
+    leases: Arc<Mutex<Vec<lease::StateRef>>>,
+    watches: Arc<Mutex<Vec<watch::StateRef>>>,
 }
 
 #[derive(Clone, Debug, serde::Serialize)]
@@ -53,81 +24,11 @@ struct Summary {
     current_timestamp: Time,
 
     #[serde(skip_serializing_if = "Vec::is_empty")]
-    leases: Vec<LeaseState>,
+    watches: Vec<watch::WatchSummary>,
+
+    #[cfg(feature = "lease")]
     #[serde(skip_serializing_if = "Vec::is_empty")]
-    watches: Vec<WatchSummary>,
-}
-
-#[derive(Clone, Debug, serde::Serialize)]
-#[serde(rename_all = "camelCase")]
-struct WatchSummary {
-    api_url: String,
-    label_selector: String,
-    #[serde(flatten)]
-    stats: WatchStats,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    checksum: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    resources: Option<Vec<Resource>>,
-}
-
-#[derive(Clone, Debug, serde::Serialize)]
-#[serde(rename_all = "camelCase")]
-struct WatchStats {
-    creation_timestamp: Time,
-
-    errors: u64,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    last_error: Option<WatchError>,
-
-    resets: u64,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    last_reset_timestamp: Option<Time>,
-
-    applies: u64,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    last_apply_timestamp: Option<Time>,
-
-    deletes: u64,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    last_delete_timestamp: Option<Time>,
-}
-
-#[derive(Clone, Debug, serde::Serialize)]
-struct WatchError {
-    message: String,
-    timestamp: Time,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-struct ObjRef {
-    kind: String,
-    api_version: String,
-    namespace: Option<String>,
-    name: Option<String>,
-    uid: Option<String>,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, serde::Serialize)]
-#[serde(rename_all = "camelCase")]
-struct Resource {
-    creation_timestamp: Option<Time>,
-    uid: String,
-    name: String,
-    namespace: String,
-    generation: Option<i64>,
-    resource_version: String,
-}
-
-#[derive(Clone, Debug, serde::Serialize)]
-#[serde(rename_all = "camelCase")]
-struct LeaseStats {
-    creation_timestamp: Time,
-
-    updates: u64,
-
-    #[serde(skip_serializing_if = "Option::is_none")]
-    last_update_timestamp: Option<Time>,
+    leases: Vec<lease::LeaseState>,
 }
 
 // === impl Diagnostics ===
@@ -150,65 +51,15 @@ impl Diagnostics {
         T: kube_client::Resource,
         T::DynamicType: Default,
     {
-        let now = Time(chrono::Utc::now());
-        let state = Arc::new(RwLock::new(WatchState {
-            api_url: api.resource_url().to_string(),
-            label_selector: label_selector.map_or_else(Default::default, ToString::to_string),
-            known: AHashMap::new(),
-            resetting: AHashMap::new(),
-            stats: WatchStats {
-                creation_timestamp: now,
-                errors: 0,
-                last_error: None,
-                resets: 0,
-                last_reset_timestamp: None,
-                applies: 0,
-                last_apply_timestamp: None,
-                deletes: 0,
-                last_delete_timestamp: None,
-            },
-        }));
-
-        let watch = Arc::downgrade(&state);
-        self.watches.lock().push(watch);
-
-        WatchDiagnostics(state)
+        let wd = WatchDiagnostics::new(api.resource_url(), label_selector);
+        self.watches.lock().push(wd.weak());
+        wd
     }
 
-    pub(crate) fn register_lease(
-        &self,
-        crate::LeaseParams {
-            name,
-            namespace,
-            claimant,
-            lease_duration,
-            renew_grace_period,
-            field_manager,
-        }: &crate::LeaseParams,
-    ) -> LeaseDiagnostics {
-        let now = Time(chrono::Utc::now());
-        let state = Arc::new(RwLock::new(LeaseState {
-            name: name.clone(),
-            namespace: namespace.clone(),
-            claimant: claimant.clone(),
-            lease_duration_seconds: lease_duration.as_secs_f64(),
-            renew_grace_period_seconds: renew_grace_period.as_secs_f64(),
-            field_manager: field_manager.clone().unwrap_or(Cow::Borrowed(
-                crate::lease::LeaseManager::DEFAULT_FIELD_MANAGER,
-            )),
-            claim: None,
-            resource_version: None,
-            stats: LeaseStats {
-                creation_timestamp: now,
-                updates: 0,
-                last_update_timestamp: None,
-            },
-        }));
-
-        let lease = Arc::downgrade(&state);
-        self.leases.lock().push(lease);
-
-        LeaseDiagnostics(state)
+    pub(crate) fn register_lease(&self, params: &crate::LeaseParams) -> LeaseDiagnostics {
+        let ld = LeaseDiagnostics::new(params);
+        self.leases.lock().push(ld.weak());
+        ld
     }
 
     pub(super) fn handle(&self, client_addr: SocketAddr, req: super::Request) -> super::Response {
@@ -255,7 +106,7 @@ impl Diagnostics {
 
     /// Collect the summaries of the remaining watches, with their resources
     /// sorted by creation.
-    fn summarize_watches(&self, with_resources: bool) -> Vec<WatchSummary> {
+    fn summarize_watches(&self, with_resources: bool) -> Vec<watch::WatchSummary> {
         let mut refs = self.watches.lock();
         // Clean up any dead weak refs, i.e. of watches that have been dropped.
         refs.retain(|w| w.upgrade().is_some());
@@ -263,33 +114,12 @@ impl Diagnostics {
             .filter_map(|wref| {
                 let watch = wref.upgrade()?;
                 let state = watch.read();
-
-                let mut resources = state.known.values().cloned().collect::<Vec<_>>();
-                resources.sort_by_key(|meta| meta.creation_timestamp.as_ref().map(|Time(t)| *t));
-
-                let checksum = if resources.is_empty() {
-                    None
-                } else {
-                    Some(checksum(&resources))
-                };
-                let resources = if with_resources {
-                    Some(resources)
-                } else {
-                    None
-                };
-
-                Some(WatchSummary {
-                    api_url: state.api_url.clone(),
-                    label_selector: state.label_selector.clone(),
-                    stats: state.stats.clone(),
-                    resources,
-                    checksum,
-                })
+                Some(state.summary(with_resources))
             })
             .collect()
     }
 
-    fn summarize_leases(&self) -> Vec<LeaseState> {
+    fn summarize_leases(&self) -> Vec<lease::LeaseState> {
         let mut refs = self.leases.lock();
         // Clean up any dead weak refs, i.e. of leases that have been dropped.
         refs.retain(|w| w.upgrade().is_some());
@@ -300,129 +130,5 @@ impl Diagnostics {
                 Some(state.clone())
             })
             .collect()
-    }
-}
-
-// === impl WatchDiagnostics ===
-
-impl WatchDiagnostics {
-    pub(crate) fn inspect<T>(&self, event: &watcher::Result<watcher::Event<T>>)
-    where
-        T: kube_client::Resource,
-        T::DynamicType: Default,
-    {
-        let to_key = |meta: &ObjectMeta| ObjRef {
-            kind: T::kind(&Default::default()).to_string(),
-            api_version: T::api_version(&Default::default()).to_string(),
-            namespace: meta.namespace.clone(),
-            name: meta.name.clone(),
-            uid: meta.uid.clone(),
-        };
-
-        // We store a summarized version fo resources to avoid storing, for
-        // example, all state for a cluster. We store only the metadata that we
-        // can use to establish a comparison between multiple controller
-        // instances and the kubernets API state.
-        let to_resource = |meta: &ObjectMeta| Resource {
-            creation_timestamp: meta.creation_timestamp.clone(),
-            name: meta.name.clone().unwrap_or_default(),
-            namespace: meta.namespace.clone().unwrap_or_default(),
-            resource_version: meta.resource_version.clone().unwrap_or_default(),
-            generation: meta.generation,
-            uid: meta.uid.clone().unwrap_or_default(),
-        };
-
-        let now = Time(chrono::Utc::now());
-        let WatchState {
-            ref mut known,
-            ref mut resetting,
-            ref mut stats,
-            ..
-        } = *self.0.write();
-        match event {
-            Ok(watcher::Event::Init) => {
-                resetting.clear();
-            }
-            Ok(watcher::Event::InitApply(res)) => {
-                resetting.insert(to_key(res.meta()), to_resource(res.meta()));
-            }
-            Ok(watcher::Event::InitDone) => {
-                std::mem::swap(known, resetting);
-                stats.resets += 1;
-                stats.last_reset_timestamp = Some(now);
-            }
-            Ok(watcher::Event::Apply(res)) => {
-                known.insert(to_key(res.meta()), to_resource(res.meta()));
-                stats.applies += 1;
-                stats.last_apply_timestamp = Some(now);
-            }
-            Ok(watcher::Event::Delete(res)) => {
-                known.remove(&to_key(res.meta()));
-                stats.deletes += 1;
-                stats.last_delete_timestamp = Some(now);
-            }
-            Err(error) => {
-                stats.errors += 1;
-                stats.last_error = Some(WatchError {
-                    message: error.to_string(),
-                    timestamp: now,
-                });
-            }
-        }
-    }
-}
-
-// === impl Resource ===
-
-impl std::hash::Hash for Resource {
-    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-        self.creation_timestamp
-            .as_ref()
-            .map(|Time(ct)| ct)
-            .hash(state);
-        self.name.hash(state);
-        self.namespace.hash(state);
-        self.resource_version.hash(state);
-        self.uid.hash(state);
-    }
-}
-
-/// Compute a SHA256 checksum of a hashable object.
-fn checksum<T: std::hash::Hash>(obj: &T) -> String {
-    use sha2::{Digest, Sha256};
-    struct Sha256Hasher(Sha256);
-    impl std::hash::Hasher for Sha256Hasher {
-        fn finish(&self) -> u64 {
-            unimplemented!("SHA-256 output is larger than u64");
-        }
-        fn write(&mut self, bytes: &[u8]) {
-            self.0.update(bytes);
-        }
-    }
-    let mut hasher = Sha256Hasher(Sha256::new());
-    obj.hash(&mut hasher);
-    format!("sha256:{:x}", hasher.0.finalize())
-}
-
-// === impl LeaseDiagnostics ===
-
-#[cfg(feature = "lease")]
-impl LeaseDiagnostics {
-    pub(crate) fn inspect(
-        &self,
-        claim: Option<Arc<crate::lease::Claim>>,
-        resource_version: String,
-    ) {
-        let mut state = self.0.write();
-        if claim.as_deref() == state.claim.as_ref()
-            && Some(&*resource_version) == state.resource_version.as_deref()
-        {
-            return;
-        }
-        let now = Time(chrono::Utc::now());
-        state.claim = claim.as_deref().cloned();
-        state.resource_version = Some(resource_version);
-        state.stats.updates += 1;
-        state.stats.last_update_timestamp = Some(now);
     }
 }
